benchmarking/pyproject.toml
[project]
name = "benchmarking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.28.0",
]
